+++ kifi-backend/modules/abook/app/com/keepit/abook/ABookController.scala
@@ -53,6 +53,7 @@ class ABookController @Inject() (
   contactRepo:ContactRepo,
   econtactRepo:EContactRepo,
   contactInfoRepo:ContactInfoRepo,
+  abookCommander:ABookCommander,
   contactsUpdater:ContactsUpdaterPlugin
 ) extends WebsiteController(actionAuthenticator) with ABookServiceController {
 
@@ -87,7 +88,7 @@ class ABookController @Inject() (
 
           val abookUpload = Json.obj("origin" -> "gmail", "ownerId" -> gUserInfo.id, "contacts" -> jsArrays(0))
           log.info(Json.prettyPrint(abookUpload))
-          val abookInfo = processUpload(userId, ABookOrigins.GMAIL, Some(gUserInfo), abookUpload)
+          val abookInfo = abookCommander.processUpload(userId, ABookOrigins.GMAIL, Some(gUserInfo), abookUpload)
           Ok(Json.toJson(abookInfo))
         } else {
           BadRequest(s"Failed to retrieve gmail contacts. Contacts API response: ${contactsResp}")
@@ -102,19 +103,10 @@ class ABookController @Inject() (
     }
   }
 
-  private def toS3Key(userId:Id[User], origin:ABookOriginType, abookOwnerInfo:Option[ABookOwnerInfo]):String = {
-    val k = s"${userId.id}_${origin.name}"
-    val ownerId = for (abookOwner <- abookOwnerInfo; ownerId <- abookOwner.id) yield ownerId
-    ownerId match {
-      case Some(id) => s"${k}_${id}"
-      case None => k
-    }
-  }
-
   def upload(userId:Id[User], origin:ABookOriginType) = Action(parse.json(maxLength = 1024 * 50000)) { request =>
     val json : JsValue = request.body
     val abookRepoEntryF: Future[ABookInfo] = Future {
-      processUpload(userId, origin, None, json)
+      abookCommander.processUpload(userId, origin, None, json)
     }
     Async {
       abookRepoEntryF.map(e => Ok(Json.toJson(e)))
@@ -130,7 +122,7 @@ class ABookController @Inject() (
     log.info(s"[upload($userId, $origin)] jsonFile=$jsonFile jsonSrc=$jsonSrc")
     val json = Json.parse(jsonSrc) // for testing
     log.info(s"[uploadJson] json=${Json.prettyPrint(json)}")
-    val abookInfoRepoEntry = processUpload(userId, origin, None, json)
+    val abookInfoRepoEntry = abookCommander.processUpload(userId, origin, None, json)
     Ok(Json.toJson(abookInfoRepoEntry))
   }
 
@@ -138,79 +130,24 @@ class ABookController @Inject() (
   def uploadJsonDirect(userId:Id[User], origin:ABookOriginType) = Action(parse.json(maxLength = 1024 * 50000)) { request =>
     val json = request.body
     log.info(s"[uploadJsonDirect($userId,$origin)] json=${Json.prettyPrint(json)}")
-    val abookInfoRepoEntry = processUpload(userId, origin, None, json)
+    val abookInfoRepoEntry = abookCommander.processUpload(userId, origin, None, json)
     Ok(Json.toJson(abookInfoRepoEntry))
   }
 
-  // shared
-  private[abook] def processUpload(userId: Id[User], origin: ABookOriginType, ownerInfoOpt:Option[ABookOwnerInfo], json: JsValue): ABookInfo = {
-    val abookRawInfoRes = Json.fromJson[ABookRawInfo](json)
-    val abookRawInfo = abookRawInfoRes.getOrElse(throw new Exception(s"Cannot parse ${json}"))
-
-    val s3Key = toS3Key(userId, origin, ownerInfoOpt)
-    s3 += (s3Key -> abookRawInfo)
-    log.info(s"[upload($userId,$origin)] s3Key=$s3Key rawInfo=$abookRawInfo}")
-
-    val abookInfoEntry = db.readWrite { implicit session =>
-      val (abookInfo, entryOpt) = origin match {
-        case ABookOrigins.IOS => { // no ownerInfo -- revisit later
-          val abookInfo = ABookInfo(userId = userId, origin = abookRawInfo.origin, rawInfoLoc = Some(s3Key), state = ABookInfoStates.PENDING)
-          val entryOpt = {
-            val s = abookInfoRepo.findByUserIdAndOrigin(userId, origin)
-            if (s.isEmpty) None else Some(s(0))
-          }
-          (abookInfo, entryOpt)
-        }
-        case ABookOrigins.GMAIL => {
-          val ownerInfo = ownerInfoOpt.getOrElse(throw new IllegalArgumentException("Owner info not set for $userId and $origin"))
-          val abookInfo = ABookInfo(userId = userId, origin = abookRawInfo.origin, ownerId = ownerInfo.id, ownerEmail = ownerInfo.email, rawInfoLoc = Some(s3Key), state = ABookInfoStates.PENDING)
-          val entryOpt = abookInfoRepo.findByUserIdOriginAndOwnerId(userId, origin, abookInfo.ownerId)
-          (abookInfo, entryOpt)
-        }
-      }
-      val entry = entryOpt match {
-        case Some(oldVal) => {
-          log.info(s"[upload($userId,$origin)] current entry: $oldVal")
-          oldVal
-        }
-        case None => abookInfoRepo.save(abookInfo)
-      }
-      entry
-    }
-    contactsUpdater.asyncProcessContacts(userId, origin, abookInfoEntry, s3Key, WeakReference(json))
-    log.info(s"[upload($userId,$origin)] created abookEntry: $abookInfoEntry")
-    abookInfoEntry
-  }
-
-
   def getContacts(userId:Id[User], maxRows:Int) = Action { request =>
     val resF:Future[JsValue] = Future {
-      getContactsDirect(userId, maxRows)
+      abookCommander.getContactsDirect(userId, maxRows)
     }
     val async: AsyncResult = Async {
       resF.map { js => Ok(js) }
     }
     async
-  }
-
-  def getContactsDirect(userId: Id[User], maxRows: Int): JsArray = {
-    val ts = System.currentTimeMillis
-    val jsonBuilder = mutable.ArrayBuilder.make[JsValue]
-    db.readOnly {
-      implicit session =>
-        contactRepo.getByUserIdIter(userId, maxRows).foreach {
-          jsonBuilder += Json.toJson(_)
-        } // TODO: paging & caching
-    }
-    val contacts = jsonBuilder.result
-    log.info(s"[getContacts($userId, $maxRows)] # of contacts returned: ${contacts.length} time-lapsed: ${System.currentTimeMillis - ts}")
-    JsArray(contacts)
   }
 
   def getEContactById(contactId:Id[EContact]) = Action { request =>
   // todo: parse email
     val resF:Future[Option[JsValue]] = Future {
-      getEContactByIdDirect(contactId)
+      abookCommander.getEContactByIdDirect(contactId)
     }
     Async {
       resF.map{ jsOpt =>
@@ -222,18 +159,11 @@ class ABookController @Inject() (
     }
   }
 
-  def getEContactByIdDirect(contactId:Id[EContact]):Option[JsValue] = {
-    val econtactOpt = db.readOnly { implicit s =>
-      econtactRepo.getById(contactId)
-    }
-    log.info(s"[getEContactByIdDirect($contactId)] res=$econtactOpt")
-    econtactOpt map { Json.toJson(_) }
-  }
 
   def getEContactByEmail(userId:Id[User], email:String) = Action { request =>
     // todo: parse email
     val resF:Future[Option[JsValue]] = Future {
-      getEContactByEmailDirect(userId, email)
+      abookCommander.getEContactByEmailDirect(userId, email)
     }
     Async {
       resF.map{ jsOpt =>
@@ -245,36 +175,15 @@ class ABookController @Inject() (
     }
   }
 
-  def getEContactByEmailDirect(userId:Id[User], email:String):Option[JsValue] = {
-    val econtactOpt = db.readOnly { implicit s =>
-      econtactRepo.getByUserIdAndEmail(userId, email)
-    }
-    log.info(s"[getEContactDirect($userId,$email)] res=$econtactOpt")
-    econtactOpt map { Json.toJson(_) }
-  }
 
   def getEContacts(userId:Id[User], maxRows:Int) = Action { request =>
     val resF:Future[JsValue] = Future {
-      getEContactsDirect(userId, maxRows)
+      abookCommander.getEContactsDirect(userId, maxRows)
     }
     val async: AsyncResult = Async {
       resF.map { js => Ok(js) }
     }
     async
-  }
-
-  def getEContactsDirect(userId: Id[User], maxRows: Int): JsArray = {
-    val ts = System.currentTimeMillis
-    val jsonBuilder = mutable.ArrayBuilder.make[JsValue]
-    db.readOnly {
-      implicit session =>
-        econtactRepo.getByUserIdIter(userId, maxRows).foreach {
-          jsonBuilder += Json.toJson(_)
-        } // TODO: paging & caching
-    }
-    val contacts = jsonBuilder.result
-    log.info(s"[getEContacts($userId, $maxRows)] # of contacts returned: ${contacts.length} time-lapsed: ${System.currentTimeMillis - ts}")
-    JsArray(contacts)
   }
 
   def getContactInfos(userId:Id[User], maxRows:Int) = Action { request =>
@@ -363,28 +272,11 @@ class ABookController @Inject() (
 
   def getABookRawInfos(userId:Id[User]) = Action { request =>
     val resF:Future[JsValue] = Future {
-      getABookRawInfosDirect(userId)
+      abookCommander.getABookRawInfosDirect(userId)
     }
     Async {
       resF.map(js => Ok(js))
     }
-  }
-
-  private [abook] def getABookRawInfosDirect(userId: Id[User]): JsValue = {
-    val abookInfos = db.readOnly {
-      implicit session =>
-        abookInfoRepo.findByUserId(userId)
-    }
-    val abookRawInfos = abookInfos.foldLeft(Seq.empty[ABookRawInfo]) {
-      (a, c) =>
-        a ++ {
-          for {k <- c.rawInfoLoc
-               v <- s3.get(k)} yield v
-        }
-    }
-    val json = Json.toJson(abookRawInfos)
-    log.info(s"[getContactsRawInfo(${userId})=$abookRawInfos json=$json")
-    json
   }
 
   def getABookInfos(userId:Id[User]) = Action { request =>
@@ -409,7 +301,7 @@ class ABookController @Inject() (
         abooks.map{ abookInfo =>
           val key = abookInfo.rawInfoLoc.getOrElse(
             origin match {
-              case ABookOrigins.IOS => toS3Key(userId, origin, None) // only ok for IOS
+              case ABookOrigins.IOS => abookCommander.toS3Key(userId, origin, None) // only ok for IOS
               case _ => throw new IllegalStateException(s"[getContactsRawInfo($userId, $origin)] rawInfoLoc not set for $abookInfo")
             }
           )
